+++ src/main/java/org/programacion/avanzada/bookstoreapp/repository/PurchaseOrderRepository.java
@@ -4,6 +4,6 @@ import org.programacion.avanzada.bookstoreapp.model.PurchaseOrder;
 import org.springframework.data.repository.CrudRepository;
 import java.util.List;
 
-public interface PurchaseOrderRepository extends CrudRepository<PurchaseOrder, Long> {
-    List<PurchaseOrder> findByCustomerId(Long customerId);
+public interface PurchaseOrderRepository extends CrudRepository<PurchaseOrder, Integer> {
+    List<PurchaseOrder> findByCustomerId(Integer customerId);
 }
